Module rocket_contrib::json

Automatic JSON (de)serialization support.

See the Json type for further details.

Enabling

This module is only available when the json feature is enabled. Enable it in Cargo.toml as follows:

[dependencies.rocket_contrib]
version = "0.4.5"
default-features = false
features = ["json"]

Structs

Json
The JSON type: implements FromData and Responder, allowing you to easily consume and respond with JSON.

JsonValue
An arbitrary JSON value.

Enums

JsonError
An error returned by the Json data guard when incoming data fails to serialize as JSON.