Struct rocket_contrib::json::Json

pub struct Json<T>(pub T);

The JSON type: implements FromData and Responder, allowing you to easily consume and respond with JSON.

Receiving JSON

If you're receiving JSON data, simply add a data parameter to your route arguments and ensure the type of the parameter is a Json<T>, where T is some type you'd like to parse from JSON. T must implement Deserialize or from serde. The data is parsed from the HTTP request body.

use rocket_contrib::json::Json;

#[post("/users", format = "json", data = "<user>")]
fn new_user(user: Json<User>) {
    /* ... */
}

You don't need to use format = "json", but it may be what you want. Using format = json means that any request that doesn't specify "application/json" as its Content-Type header value will not be routed to the handler.

Sending JSON

If you're responding with JSON data, return a Json<T> type, where T implements Serialize from serde. The content type of the response is set to application/json automatically.

use rocket_contrib::json::Json;

#[get("/users/<id>")]
fn user(id: usize) -> Json<User> {
    let user_from_id = User::from(id);
    /* ... */
    Json(user_from_id)
}

Incoming Data Limits

The default size limit for incoming JSON data is 1MiB. Setting a limit protects your application from denial of service (DoS) attacks and from resource exhaustion through high memory consumption. The limit can be increased by setting the limits.json configuration parameter. For instance, to increase the JSON limit to 5MiB for all environments, you may add the following to your Rocket.toml:

[global.limits]
json = 5242880

Methods

impl<T> Json<T>

pub fn into_inner(self) -> T

Consumes the JSON wrapper and returns the wrapped item.

Example

let string = "Hello".to_string();
let my_json = Json(string);
assert_eq!(my_json.into_inner(), "Hello".to_string());