Struct rocket_contrib::json::JsonValue

pub struct JsonValue(pub Value);

An arbitrary JSON value.

This structure wraps serde's Value type. Importantly, unlike Value, this type implements Responder, allowing a value of this type to be returned directly from a handler.

Responder

The Responder implementation for JsonValue serializes the represented value into a JSON string and sets the string as the body of a fixed-sized response with a Content-Type of application/json.

Usage

A value of this type is constructed via the json! macro. The macro and this type are typically used to construct JSON values in an ad-hoc fashion during request handling. This looks something like:

use rocket_contrib::json::JsonValue;

#[get("/json")]
fn get_json() -> JsonValue {
    json!({
        "id": 83,
        "values": [1, 2, 3, 4]
    })
}