Struct rocket_contrib::Template

pub struct Template(_, _);

The Template type implements generic support for template rendering in Rocket.

Templating in Rocket words by first discovering all of the templates inside the template directory. The template directory is configurable via the template_dir configuration parameter and defaults to templates/. The path set in template_dir should be relative to the Rocket configuration file. See the configuration chapter of the guide for more information on configuration.

Templates are discovered according to their extension. At present, this library supports the following templates and extensions:

Tera: .tera

Handlebars: .hbs

Any file that ends with one of these extension will be discovered and rendered with the corresponding templating engine. The name of the template will be the path to the template file relative to template_dir minus at most two extensions. The following are examples of template names (on the right) given that the template is at the path on the left.

{template_dir}/index.html.hbs => index

{template_dir}/index.tera => index

{template_dir}/index.hbs => index

{template_dir}/dir/index.hbs => dir/index

{template_dir}/dir/index.html.tera => dir/index

{template_dir}/index.template.html.hbs => index.template

{template_dir}/subdir/index.template.html.hbs => subdir/index.template

The recommended naming scheme is to use two extensions: one for the file type, and one for the template extension. This means that template extensions should look like: .html.hbs, .html.tera, .xml.hbs, etc.

Templates are rendered with the render method. The method takes in the name of a template and a context to render the template with. The context can be any type that implements Serialize from Serde and would serialize to an Object value.

Usage

To use, add the handlebars_templates feature, the tera_templates feature, or both, to the rocket_contrib dependencies section of your Cargo.toml:

[dependencies.rocket_contrib]
version = "*"
default-features = false
features = ["handlebars_templates", "tera_templates"]

The Template type implements Rocket's Responder trait, so it can be returned from a request handler directly:

#[get("/")]
fn index() -> Template {
    let context = ...;
    Template::render("index", &context)
}

Methods

impl Template

fn render<S, T>(name: S, context: &T) -> Template where S: AsRef<str>, T: Serialize

Render the template named name with the context context. The context can be of any type that implements Serialize. This is typically a HashMap or a custom struct.

Example

use std::collections::HashMap;
use rocket_contrib::Template;

// Create a `context`. Here, just an empty `HashMap`.
let mut context = HashMap::new();

let template = Template::render("index", &context);