Struct rocket_contrib::serve::StaticFiles

pub struct StaticFiles { /* fields omitted */ }

Custom handler for serving static files.

This handler makes it simple to serve static files from a directory on the local file system. To use it, construct a StaticFiles using either StaticFiles::from() or StaticFiles::new() then simply mount the handler at a desired path.

Options

The handler's functionality can be customized by passing an Options to StaticFiles::new().

Example

To serve files from the /static directory at the /public path, allowing index.html files to be used to respond to requests for a directory (the default), you might write the following:

use rocket_contrib::serve::StaticFiles;

fn main() {
    rocket::ignite()
        .mount("/public", StaticFiles::from("/static"))
        .launch();
}

With this set-up, requests for files at /public/<path..> will be handled by returning the contents of /static/<path..>. Requests for directories at /public/<directory> will be handled by returning the contents of /static/<directory>/index.html.

If your static files are stored relative to your crate and your project is managed by Cargo, you should either use a relative path and ensure that your server is started in the crate's root directory or use the CARGO_MANIFEST_DIR to create an absolute path relative to your crate root. For example, to serve files in the static subdirectory of your crate at /, you might write:

use rocket_contrib::serve::StaticFiles;

fn main() {
    rocket::ignite()
        .mount("/", StaticFiles::from(concat!(env!("CARGO_MANIFEST_DIR"), "/static")))
        .launch();
}

Methods

impl StaticFiles

pub fn from<P: AsRef<Path>>(path: P) -> Self

Constructs a new StaticFiles that serves files from the file system path. By default, Options::Index is enabled. To serve static files with other options, use StaticFiles::new().

Example

Serve the static files in the /www/public local directory on path /static.

use rocket_contrib::serve::StaticFiles;

fn main() {
    rocket::ignite()
        .mount("/static", StaticFiles::from("/www/public"))
        .launch();
}

pub fn new<P: AsRef<Path>>(path: P, options: Options) -> Self

Constructs a new StaticFiles that serves files from the file system path with options enabled.

Example

Serve the static files in the /www/public local directory on path /static without serving index files or dot files. Additionally, serve the same files on /pub while also seriving index files and dot files.

use rocket_contrib::serve::{StaticFiles, Options};

fn main() {
    let options = Options::Index | Options::DotFiles;
    rocket::ignite()
        .mount("/static", StaticFiles::from("/www/public"))
        .mount("/pub", StaticFiles::new("/www/public", options))
        .launch();
}